Shifted ActorPool to actor_pool.rs file
